Trade Slider working with build issues resolved
