+++ src/components/general/MarketOrderAction/stories.tsx
@@ -8,4 +8,4 @@ component: MarketOrderAction
 } as Meta
 
 export const Default: Story = (args) =>
-<MarketOrderAction {...args} />
+<MarketOrderAction {...args} setOpenOrder={args.setOpenOrder} price={args.price} amount={args.amount} setPrice={args.setPrice} setAmount={args.setAmount} />
