+++ src/templates/dashboard/blocks/MarketOrder/stories.tsx
@@ -8,4 +8,4 @@ component: MarketOrder
 } as Meta
 
 export const Default: Story = (args) =>
-<MarketOrder {...args} />
+<MarketOrder {...args} setOpenOrder={args.setOpenOrder} price={args.price} amount={args.amount} setPrice={args.setPrice} setAmount={args.setAmount} />
