+++ src/templates/dashboard/blocks/Transactions/stories.tsx
@@ -8,4 +8,4 @@ component: Transactions
 } as Meta
 
 export const Default: Story = (args) =>
-<Transactions {...args} />
+  <Transactions {...args} data={args.data} openOrderData={args.openOrderData} newTradeData={args.newTradeData} remove={args.remove} />
