+++ src/templates/landing/styles.ts
